Import of BRADFITZ/Brackup-1.10 from CPAN.
gitpan-cpan-distribution: Brackup
gitpan-cpan-version:      1.10
gitpan-cpan-path:         BRADFITZ/Brackup-1.10.tar.gz
gitpan-cpan-author:       BRADFITZ
gitpan-cpan-maturity:     released

+++ lib/Brackup/Util.pm
@@ -5,10 +5,12 @@ require Exporter;
 
 use vars qw(@ISA @EXPORT_OK);
 @ISA = ('Exporter');
-@EXPORT_OK = qw(tempfile tempdir slurp valid_params noclobber_filename);
+@EXPORT_OK = qw(tempfile tempfile_obj tempdir slurp valid_params noclobber_filename io_print_to_fh io_sha1);
 
 use File::Path qw();
 use Carp;
+use Fcntl qw(O_RDONLY);
+use Digest::SHA1;
 
 my $mainpid = $$;
 my $_temp_directory;
@@ -16,7 +18,7 @@ my $_temp_directory;
 END {
     # will happen after File::Temp's cleanup
     if ($$ == $mainpid and $_temp_directory) {
-        File::Path::rmtree($_temp_directory, 0, 1);
+        File::Path::rmtree($_temp_directory, 0, 1) unless $ENV{BRACKUP_TEST_NOCLEANUP};
     }
 }
 use File::Temp ();
@@ -25,7 +27,7 @@ sub _get_temp_directory {
     # Create temporary directory if we need one. By default, all temporary
     # files will be placed in it.
     unless (defined($_temp_directory)) {
-        $_temp_directory = File::Temp::tempdir(CLEANUP => 1);
+        $_temp_directory = File::Temp::tempdir(CLEANUP => $ENV{BRACKUP_TEST_NOCLEANUP} ? 0 : 1);
     }
 
     return $_temp_directory;
@@ -34,6 +36,10 @@ sub _get_temp_directory {
 sub tempfile {
     my (@ret) = File::Temp::tempfile(DIR => _get_temp_directory());
     return wantarray ? @ret : $ret[0];
+}
+
+sub tempfile_obj {
+    return File::Temp->new(DIR => _get_temp_directory(), CLEANUP => $ENV{BRACKUP_TEST_NOCLEANUP} ? 0 : 1);
 }
 
 # Utils::tempdir() accepts the same options as File::Temp::tempdir.
@@ -45,8 +51,15 @@ sub tempdir {
 
 sub slurp {
     my $file = shift;
-    open(my $fh, $file) or die "Failed to open $file: $!\n";
-    return do { local $/; <$fh>; }
+    my %opts = @_;
+    my $fh;
+    if ($opts{decompress} and eval { require IO::Uncompress::AnyUncompress }) {
+        $fh = IO::Uncompress::AnyUncompress->new($file)
+            or die "Failed to open file $file: $IO::Uncompress::AnyUncompress::AnyUncompressError";
+    } else {
+        sysopen($fh, $file, O_RDONLY) or die "Failed to open $file: $!";
+    }
+    return do { local $/; <$fh>; };
 }
 
 sub valid_params {
@@ -64,6 +77,35 @@ sub noclobber_filename {
     for (my $i = 1; ; $i++) {
         return "$filename.$i" if ! -e "$filename.$i";
     }
+}
+
+# Prints all data from an IO::Handle to a filehandle
+sub io_print_to_fh {
+    my ($io_handle, $fh, $sha1) = @_;
+    my $buf;
+    my $bytes = 0;
+
+    while($io_handle->read($buf, 4096)) {
+        print $fh $buf;
+        $bytes += length $buf;
+        $sha1->add($buf) if $sha1;
+    }
+
+    return $bytes;
+}
+
+# computes sha1 of data in an IO::Handle
+sub io_sha1 {
+    my ($io_handle) = @_;
+    
+    my $sha1 = Digest::SHA1->new;
+    my $buf;
+    
+    while($io_handle->read($buf, 4096)) {
+        $sha1->add($buf);
+    }
+
+    return $sha1->hexdigest;
 }
 
 1;
